small edits on stack and queue to complete  doc strings
